refactor(rollback): extract deepEqual with deliberate conflict-equality contract
Move deepEqual into services/api/rollback/deep-equal.ts and fix two accidental
=== semantics deliberately (called out in PR):
- NaN equals NaN (Object.is-style) so a NaN-valued field cannot perpetually
  conflict; signed zero left on === so numeric zero never newly conflicts.
- Date vs string compared by instant (parse then getTime) instead of
  toISOString()-vs-String(), so e.g. permission expiresAt no longer spuriously
  conflicts on a non-canonical serialization.
Object key presence remains significant (undefined field != missing key).
100% branch coverage; existing rollback-service + ai-undo suites pass unchanged;
web typecheck green.
